feat: Add photo retrieval and update functionality for students
This commit adds the `getPhoto` and `updatePhoto` methods to the `EstudianteController` class. The `getPhoto` method retrieves the photo of a student based on their ID, while the `updatePhoto` method updates the photo of a student with a new file. These methods are used to handle the retrieval and updating of student photos.

The first part of entity  "notifications" (DAO and CONTROLLER) was implemented, the model will be shown later, for reasons of the Observer pattern.......