feat: Add functionality to change student password
Ya funciona el cambiar contraseña.

por completar aqui:
- Validad que esta opción solo la tenga el usuario tipo 5 (estudiante.)

This commit adds a new page for changing the student's password. It obtains the ID of the logged-in user and sends the ID and the confirmed password in the data object to the server. The server then updates the password in the database. This feature improves the security and user experience of the application.